+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## Unreleased
+## [0.23.0] - 2024-11-17
 
 ### Changed
 
@@ -223,6 +223,7 @@
 
 - Initial release
 
+[0.23.0]: https://github.com/tinted-theming/tinty/compare/v0.22.0...v0.23.0
 [0.22.0]: https://github.com/tinted-theming/tinty/compare/v0.21.1...v0.22.0
 [0.21.1]: https://github.com/tinted-theming/tinty/compare/v0.21.0...v0.21.1
 [0.21.0]: https://github.com/tinted-theming/tinty/compare/v0.20.1...v0.21.0